+++ fantasy/project.yaml
@@ -11,3 +11,23 @@ default_db: warehouse
 parameters:
   base_url:  '' #no prefix for prod
   api_key: ''
+  teams:
+    - 15
+    - 39
+    - 1883502
+    - 7390454
+    - 5
+    - 8214850
+    - 8255756
+    - 2586976
+    - 7119388
+    - 726228
+    - 1838315
+    - 6209166
+    - 111474
+    - 8254400
+    - 7391077
+    - 8260983
+    - 8204512
+    - 350190
+
